webapp/server/requirements.txt
# Django Backend Requirements for Real Estate API
django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-filter==23.3

# Optional: Environment management (không cần vì dùng os.getenv)
# python-decouple==3.8
# dj-database-url==2.1.0

# ML inference and data processing
joblib==1.3.2
scikit-learn==1.3.2
pandas==2.1.1
numpy==1.24.3
pyarrow==14.0.1

# HDFS integration for model loading
hdfs==2.7.0
pyspark==3.5.0

# HTTP client for external API calls
httpx==0.25.0

# Additional utilities
python-dateutil==2.8.2

# ML models (optional - will be handled gracefully if not available)
# xgboost==1.7.6
# lightgbm==4.1.0
# catboost==1.2.2

# Additional dependencies for analytics and processing
scipy==1.11.4
# lightgbm==4.1.0
# catboost==1.2.2
